requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2==2.9.5
asgiref==3.5.2
django-environ==0.9.0
sqlparse==0.4.3
tzdata==2022.6
Pillow==9.3.0
djangorestframework_simplejwt==5.2.2
django-filter==22.1
Faker==15.3.4
factory-boy==3.2.1